add 1 migration to project
