+++ docs/api_schema.md
@@ -1,0 +1,21 @@
+# Стандарты наименования API и структуры ответов
+
+## Наименование API точек и выбор HTTP метола
+
+### Основные принципы
+- Используйте **snake_case** для наименования параметров.
+- Используйте **kebab-case** для наименования эндпоинтов.
+- При нейминге опираемся на REST API best practice naming.
+- Все API точки должны заканчиваться /
+- Можно оставлять один из PUT/PATCH в зависимости от того какая логика требуется от CRUD.
+
+
+### Примеры REST нейминга:
+|HTTP метод|Название точки|Описание действия|
+|-|-|-|
+|**GET**|`/api/articles/`| Получение списка статей.|
+|**POST**|`/api/articles/`| Создание новой статьи.|
+|**GET**|`/api/articles/{article_id}/`| Получение конкретной статьи по ID.|
+|**PUT**|`/api/articles/{article_id}/`| Полное обновление статьи (все поля из модели).|
+|**PATCH**|`/api/articles/{article_id}/`| Частичное обновление статьи.|
+|**DELETE**|`/api/articles/{article_id}/`| Удаление статьи.|
